Moved website higher, used Bootstrap Icon font instead of SVG template.  Added Bootstrap Icon CDN to project_index/project_page
Since I just built the website I feel it's writeup is one of the best on the site, but I feel it's the least important of the major projects, so it's below the SGA-Tool, Procedural Planet, and Univox.

Because I can't get SVG use statements to work to save my life; I relied on pystache templates to inject SVG content into the site; but I could sidestep that completely by just using the bootstrap icon CDN . This gives me the added benefit of being able to use any icon, at the cost of a superflous CSS incllude (in most cases).